Simple Regression, Multiple Regression, Pipeline
Objectives:
Develop prediction models
In this task, I'll develop several models that will predict the price of the car using the variables or features. This is just an estimate but should give us an objective idea of how much the car should cost.

Some questions I answered in this module
Do I know if the dealer if offering fair value for my trade-in?
Do I know if I put a fair value on my car?


A model will help us understand the exact relationship between different variables and how these variables are used to predict the result.